Add debugger logic to os environ config
